DZ/steam_games/requirements.txt
Django==4.2.16
python3-openid==3.2.0
requests==2.32.3